Update unit tests for Pocket class
